+++ src/redsun/device/__init__.py
@@ -8,8 +8,8 @@ from ._acquisition import (
     TriggerType,
 )
 from ._attrs import AttrR, AttrRW, AttrT, AttrW
-from ._base import Device, PDevice
-from ._soft import SoftAttrR, SoftAttrRW, SoftAttrT
+from ._base import Device, HasChildren, PDevice
+from ._soft import SoftAttr, SoftAttrR, SoftAttrRW, SoftAttrT
 
 __all__ = [
     "AcquisitionController",
@@ -21,8 +21,10 @@ __all__ = [
     "DataWriter",
     "Device",
     "FlyerController",
+    "HasChildren",
     "PDevice",
     "PrepareInfo",
+    "SoftAttr",
     "SoftAttrR",
     "SoftAttrRW",
     "SoftAttrT",